docs: consolidate agent guides
