+++ apps/web/tsconfig.json
@@ -11,7 +11,7 @@
     "module": "ESNext",
     "moduleResolution": "Bundler",
     "types": ["bun", "vite/client"],
-    "paths": { "@docs/*": ["../../docs/site/*"] },
+    "paths": { "@docs/*": ["../../docs/*"] },
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
