--- docs/README.md
@@ -1,11 +1,0 @@
-# tardigrade docs
-
-- [quickstart.md](quickstart.md): all the core concepts you need to get started.
-- [how-to/migrate.md](how-to/migrate.md): move an existing agent application and its history to tardigrade.
-- [trace-review.md](trace-review.md): the trace review and eval discovery product specification.
-- [how-to/server.md](how-to/server.md): the HTTP server.
-- [how-to/cli.md](how-to/cli.md): the CLI.
-- [how-to/celld.md](how-to/celld.md): deploy an actor project to a Celld fleet.
-- [explanations/why.md](explanations/why.md): why tardigrade exists. {transitions} = f(log), and what the log-as-state shape enables.
-- [explanations/threads.md](explanations/threads.md): actor definitions, durable threads, storage boundaries, and child placement.
-- [site/README.md](site/README.md): author pages rendered by the web application.
