+++ docs/platforms/bun.mdx
